Add menu to the main file
